day 13 done
who said it was unlucky? to be honest this one wasn't too bad, unfortunately it took me a long time because I did the first part with brute force and needed to learn a bit of z3 before I could finish part 2. Overall a pretty good question in my humble opinion.